requirements.txt
erddapy==2.2.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
PyYAML==6.0.1